Remove un-necessary abstract class
